indexing: more cleanup and adding more comments
