+++ src/javasources/KTool/src/org/kframework/backend/java/indexing/pathIndex/trie/TrieLeaf.java
@@ -6,6 +6,8 @@ import java.util.LinkedHashSet;
 import java.util.Set;
 
 /**
+ * A leaf node in the trie, i.e., one that has no children
+ * <p/>
  * Author: Owolabi Legunsen
  * 1/2/14: 7:30 PM
  */
@@ -23,28 +25,8 @@ public class TrieLeaf extends TrieNode {
         indices.add(index);
     }
 
-
     @Override
     public String toString() {
         return getValue() + " => " + getIndices();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        if (!super.equals(o)) return false;
-
-        TrieLeaf trieLeaf = (TrieLeaf) o;
-
-        return indices.equals(trieLeaf.indices);
-
-    }
-
-    @Override
-    public int hashCode() {
-        int result = super.hashCode();
-        result = 31 * result + indices.hashCode();
-        return result;
     }
 }
